Move scroll-composition helper out of __tests__ to fix CI.
Jest was treating the bare .ts helper as an empty test suite; relocate to test-utils/ and guard against regression.
